# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "GeometryView": {
        "additionalProperties": false,
        "properties": {
          "has_disputed_mass": {
            "title": "Has Disputed Mass",
            "type": "boolean"
          },
          "spec": {
            "$ref": "#/components/schemas/GeometrySpec"
          },
          "svg": {
            "title": "Svg",
            "type": "string"
          },
          "total_height_m": {
            "title": "Total Height M",
            "type": "number"
          }
        },
        "required": [
          "spec",
          "svg",
          "has_disputed_mass",
          "total_height_m"
        ],
        "title": "GeometryView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Municipal structural intelligence as an institutional agent fleet.\n\nA decision-support prototype. It delivers information and performs clerical\nexecution. It never issues tactical recommendations, evacuation orders, crew\nassignments, or fire-behaviour predictions.",
    "title": "FIRST DUE",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/buildings/{address_id}": {
      "get": {
        "operationId": "building_profile_api_v1_buildings__address_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildingProfileView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One building profile",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/buildings/{address_id}/geometry": {
      "get": {
        "description": "The spec plus its static fallback.\n\nDisputed mass is marked in the data, so the fallback shows it as disputed\ntoo -- the conflict is not a property of the renderer.",
        "operationId": "building_geometry_api_v1_buildings__address_id__geometry_get",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeometryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Renderable geometry, with an SVG fallback",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/buildings/{address_id}/surveys": {
      "get": {
        "operationId": "building_surveys_api_v1_buildings__address_id__surveys_get",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SurveyRecord"
                  },
                  "title": "Response Building Surveys Api V1 Buildings  Address Id  Surveys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company surveys recorded for this building",
        "tags": [
          "console"
        ]
      },
      "post": {
        "description": "Record what a crew saw.\n\nThis is the only input that can set ``human_verified`` and the only one that\ncan close a conflict. Attributes the crew did not observe stay open.",
        "operationId": "record_survey_api_v1_buildings__address_id__surveys_post",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveySubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Survey Api V1 Buildings  Address Id  Surveys Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a completed company survey",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/buildings/{address_id}/timeline": {
      "get": {
        "description": "Everything that ever happened to this profile, in order. Never edited.",
        "operationId": "building_timeline_api_v1_buildings__address_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "address_id",
            "required": true,
            "schema": {
              "title": "Address Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimelineEventView"
                  },
                  "title": "Response Building Timeline Api V1 Buildings  Address Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The building's append-only timeline",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/conflicts/{conflict_id}/referral": {
      "post": {
        "description": "Stage, never file.\n\nReturns ``202``: the referral exists and is waiting for a human. Filing it\naccuses a property owner of unpermitted construction, which is a captain's\ndecision and not an agent's.",
        "operationId": "stage_referral_api_v1_conflicts__conflict_id__referral_post",
        "parameters": [
          {
            "in": "path",
            "name": "conflict_id",
            "required": true,
            "schema": {
              "title": "Conflict Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Stage Referral Api V1 Conflicts  Conflict Id  Referral Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage a building-department referral for a conflict",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/districts/{district_id}/poll": {
      "post": {
        "description": "Poll every source, materialize, and re-rank.\n\nIn production a scheduler drives this; exposing it lets the console show the\nloop running and lets the demo be a single request. It is idempotent: a\nsecond call re-derives what exists and writes none of it again.\n\n``approve`` defaults to false, which leaves any staged referral waiting for\na human -- the honest default.",
        "operationId": "poll_district_api_v1_districts__district_id__poll_post",
        "parameters": [
          {
            "in": "path",
            "name": "district_id",
            "required": true,
            "schema": {
              "title": "District Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "approve",
            "required": false,
            "schema": {
              "default": false,
              "title": "Approve",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Poll District Api V1 Districts  District Id  Poll Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run one slow-loop pass over a district",
        "tags": [
          "console"
        ]
      }
    },
...
    "/api/v1/districts/{district_id}/stats": {
      "get": {
        "operationId": "district_stats_api_v1_districts__district_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "district_id",
            "required": true,
            "schema": {
              "title": "District Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictStatsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "District readiness statistics",
        "tags": [
          "console"
        ]
      }
    },
...
  }
}